+++ src/tiled/geometry.h
@@ -1,6 +1,8 @@
 /*
  * geometry.h
  * Copyright 2010-2011, Stefan Beller <stefanbeller@googlemail.com>
+ * Copyright 2017, Benjamin Trotter <bdtrotte@ucsc.edu>
+ * Copyright 2017-2023, Thorbjørn Lindeijer <bjorn@lindeijer.nl>
  *
  * This file is part of Tiled.
  *
@@ -27,12 +29,15 @@
 
 namespace Tiled {
 
-QVector<QPoint> pointsOnEllipse(int x0, int y0, int x1, int y1);
+QVector<QPoint> pointsOnEllipse(int xm, int ym, int a, int b);
 QRegion ellipseRegion(int x0, int y0, int x1, int y1);
 QVector<QPoint> pointsOnLine(int x0, int y0, int x1, int y1, bool manhattan = false);
 
-inline QVector<QPoint> pointsOnEllipse(QPoint a, QPoint b)
-{ return pointsOnEllipse(a.x(), a.y(), b.x(), b.y()); }
+inline QVector<QPoint> pointsOnEllipse(QPoint center, int radiusX, int radiusY)
+{ return pointsOnEllipse(center.x(), center.y(), radiusX, radiusY); }
+
+inline QRegion ellipseRegion(QRect rect)
+{ return ellipseRegion(rect.left(), rect.top(), rect.right(), rect.bottom()); }
 
 inline QVector<QPoint> pointsOnLine(QPoint a, QPoint b, bool manhattan = false)
 { return pointsOnLine(a.x(), a.y(), b.x(), b.y(), manhattan); }