+++ NEWS.md
@@ -9,6 +9,7 @@
 * Fixed hover highlight rendering with active parallax factor (#3669)
 * Fixed updating of object selection outlines when changing parallax factor (#3669)
 * Fixed "Offset Map" action to offset all objects when choosing "Whole Map" as bounds
+* Fixed several issues with drawing ellipses (#3776)
 * Godot 4 plugin: Export custom tile properties as Custom Data Layers (with Kevin Harrison, #3653)
 * AppImage: Updated to Sentry 0.6.4
 * Qt 6: Increased the image allocation limit from 1 GB to 4 GB (#3616)